refactor billing helpers to lib file
